# Conflicts:
#	boards.txt
#	package/package_pico_index.template.json
#	tools/makeboards.py

+++ libraries/WebServer/examples/PostServer/PostServer.ino
@@ -62,7 +62,7 @@ void handleForm() {
   } else {
     digitalWrite(led, 1);
     String message = "POST form was:\n";
-    for (uint8_t i = 0; i < server.args(); i++) {
+    for (int i = 0; i < server.args(); i++) {
       message += " " + server.argName(i) + ": " + server.arg(i) + "\n";
     }
     server.send(200, "text/plain", message);
@@ -80,7 +80,7 @@ void handleNotFound() {
   message += "\nArguments: ";
   message += server.args();
   message += "\n";
-  for (uint8_t i = 0; i < server.args(); i++) {
+  for (int i = 0; i < server.args(); i++) {
     message += " " + server.argName(i) + ": " + server.arg(i) + "\n";
   }
   server.send(404, "text/plain", message);
